requirements.txt
fintoc>=2.18.0
python-dotenv>=1.0.0
cryptography>=41.0.0
